// src/app/shared/models/enums/StorageEnum.ts
export enum StorageEnum {
  CATEGORIAS = "cuide_bem_categorias",
  FAIXAS = "cuide_bem_faixas",
  IMAGENS = "cuide_bem_imagens",
  PACOTES = "cuide_bem_pacotes",
  PRODUTOS = "cuide_bem_produtos",
  OBJETO_ENVIO = "cuide_bem_objeto_envio",
  USUARIO = "cuide_bem_usuario",
  CLIENTE = "cuide_bem_cliente",
}